refactor: sideshift effects & functions
