+++ react/lib/components/Widget/WidgetContainer.tsx
@@ -106,6 +106,7 @@ export const WidgetContainer: React.FunctionComponent<WidgetContainerProps> =
 
     const [thisPaymentId, setThisPaymentId] = useState<string | undefined>();
     const [thisPrice, setThisPrice] = useState(0);
+    const [usdPrice, setUsdPrice] = useState(0);
     useEffect(() => {
       if ((paymentId === undefined || paymentId === '') && !disablePaymentId) {
         const newPaymentId = generatePaymentId(8);
@@ -191,7 +192,9 @@ export const WidgetContainer: React.FunctionComponent<WidgetContainerProps> =
     const getPrice = useCallback(
       async () => {
         const price = await getFiatPrice(currency, to, apiBaseUrl)
+        const usdPrice = await getFiatPrice('USD', to, apiBaseUrl)
         if (price !== null) setThisPrice(price)
+        if (usdPrice !== null) setUsdPrice(usdPrice)
       }
       , [currency, to, apiBaseUrl]
     );
@@ -241,6 +244,7 @@ export const WidgetContainer: React.FunctionComponent<WidgetContainerProps> =
           setCurrencyObject={setCurrencyObj}
           randomSatoshis={randomSatoshis}
           price={thisPrice}
+          usdPrice={usdPrice}
           success={success}
           disabled={disabled}
           editable={editable}
